Switched parsing of yaml, to be in io_parameters, now acts as a required, with default values of commec-dbs, to be grabbed from here, rather than updating the file paths after search handler creation.
